[Refactor] : 입출력 구현
