--- src/main/java/org/nextstep/Main.java
@@ -1,7 +1,0 @@
-package org.nextstep;
-
-public class Main {
-    public static void main(String[] args) {
-        System.out.println("Hello world!");
-    }
-}
