+++ src/main/java/org/nextstep/domain/Calculator.java
@@ -1,4 +1,4 @@
-package org.nextstep;
+package org.nextstep.domain;
 
 public interface Calculator {
     double calculate();
